021
Before when Pool section was skipped, certain data was not collected. These functions will run separate when Pool is skipped to make sure the other sections get the required data. (JohnB)
- GatherXSPoolMemoryData
- GatherXSPoolStorageData
- GatherXSPoolNsetworkingData
Changed the following functions with the previous created functions and data, and changed the output to be more aligned to XenCenter (JohnB)
- OutputPoolNetworking
- OutputPoolStorage
- OutputPoolMemory
- OutputHostNetworking
- OutputHostStorage
- OutputHostMemory
Changed the OutputHostUpdates functions, sorting on update name (JohnB)